Factor out 'Bar' widget.
